add many modules

+++ src/app/api/categories/route.js
@@ -1,22 +1,32 @@
 import { Category } from '@/models/Category';
 import mongoose from 'mongoose';
+import { isAdmin } from '@/app/api/auth/[...nextauth]/route';
 
 export const POST = async req => {
   await mongoose.connect(process.env.MONGO_URL);
   const { name } = await req.json();
-  const categoryDoc = await Category.create({ name });
-  return Response.json({ categoryDoc });
+
+  if (await isAdmin()) {
+    const categoryDoc = await Category.create({ name });
+    return Response.json({ categoryDoc });
+  } else {
+    return Response.json({})
+  }
 };
 
 export const PUT = async req => {
   await mongoose.connect(process.env.MONGO_URL);
   const { _id, name } = await req.json();
-  await Category.updateOne({ _id }, { name });
+
+  if (await isAdmin()) {
+    await Category.updateOne({ _id }, { name });
+  }
   return Response.json(true);
 };
 
 export const GET = async () => {
   await mongoose.connect(process.env.MONGO_URL);
+
   return Response.json(await Category.find());
 };
 
@@ -24,6 +34,9 @@ export const DELETE = async req => {
   await mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
   const _id = url.searchParams.get('_id');
-  await Category.deleteOne({ _id });
+
+  if (await isAdmin()) {
+    await Category.deleteOne({ _id });
+  }
   return Response.json(true);
 };
