+++ src/app/api/menu-items/route.js
@@ -1,22 +1,32 @@
 import { MenuItem } from '@/models/MenuItem';
 import mongoose from 'mongoose';
+import { isAdmin } from '@/app/api/auth/[...nextauth]/route';
 
 export async function POST(req) {
   await mongoose.connect(process.env.MONGO_URL);
   const data = await req.json();
-  const menuItemDoc = await MenuItem.create(data);
-  return Response.json(menuItemDoc);
+
+  if (await isAdmin()) {
+    const menuItemDoc = await MenuItem.create(data);
+    return Response.json(menuItemDoc);
+  } else {
+    return Response.json({});
+  }
 }
 
 export async function PUT(req) {
   await mongoose.connect(process.env.MONGO_URL);
-  const { _id, ...data } = await req.json();
-  await MenuItem.findByIdAndUpdate(_id, data);
+
+  if (await isAdmin()) {
+    const { _id, ...data } = await req.json();
+    await MenuItem.findByIdAndUpdate(_id, data);
+  }
   return Response.json(true);
 }
 
 export async function GET() {
   await mongoose.connect(process.env.MONGO_URL);
+
   return Response.json(await MenuItem.find());
 }
 
@@ -24,6 +34,9 @@ export async function DELETE(req) {
   await mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
   const _id = url.searchParams.get('_id');
-  await MenuItem.deleteOne({ _id });
+
+  if (await isAdmin()) {
+    await MenuItem.deleteOne({ _id });
+  }
   return Response.json(true);
 }
